# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Web Scraper API",
    "version": "1.0.0",
    "description": "Documentation for the Web Scraper API"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/data": {
      "get": {
        "summary": "Retrieve Data from a Single Web Page",
        "description": "Use this endpoint to scrape and retrieve data from a single web page specified by the URL.",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "schema": {
              "type": "string"
            },
            "description": "The URL of the web page to be scraped."
          }
        ],
        "responses": {
          "200": {
            "description": "Data has been successfully retrieved from the specified web page."
          }
        },
        "example": {
          "usage": {
            "summary": "Usage Example",
            "value": "To retrieve data from the web page at https://wsa-test.vercel.app/blog/the-disappointing-reality-of-junk-food`, make a GET request to `http://localhost:5000/data?url=https://wsa-test.vercel.app/blog/the-disappointing-reality-of-junk-food`."
          }
        }
      }
    },
    "/data-from-script": {
      "get": {
        "summary": "Retrieve Data from a Single Web Page Using a Script",
        "description": "Use this endpoint to scrape data from a single web page based on the URL provided, utilizing a custom script for extraction.",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "schema": {
              "type": "string"
            },
            "description": "The URL of the web page to be scraped."
          }
        ],
        "responses": {
          "200": {
            "description": "Data has been successfully retrieved from the specified web page using the custom script."
          }
        },
        "example": {
          "usage": {
            "summary": "Usage Example",
            "value": "To retrieve data from the web page at `https://wsa-test.vercel.app/blog/the-disappointing-reality-of-junk-food` using a custom script, make a GET request to `http://localhost:5000/data-from-script?url=https://wsa-test.vercel.app/blog/the-disappointing-reality-of-junk-food`."
          }
        }
      }
    },
    "/pages": {
      "get": {
        "summary": "Retrieve Links from a Web Page",
        "description": "This endpoint enables you to fetch links from a specified web page.",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "schema": {
              "type": "string"
            },
            "description": "The URL of the web page to obtain links from."
          }
        ],
        "responses": {
          "200": {
            "description": "Links have been successfully retrieved from the specified web page."
          }
        },
        "example": {
          "usage": {
            "summary": "Usage Example",
            "value": "To retrieve links from the web page at `https://wsa-test.vercel.app`, make a GET request to `http://localhost:5000/pages?url=https://wsa-test.vercel.app`."
          }
        }
      }
    },
    "/getAll": {
      "get": {
        "summary": "Retrieve Data from Multiple Web Pages",
        "description": "Use this endpoint to collect data from multiple web pages linked from the specified URL.",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "schema": {
              "type": "string"
            },
            "description": "The URL of the web page to start collecting data from."
          }
        ],
        "responses": {
          "200": {
            "description": "Data has been successfully retrieved from multiple web pages."
          }
        },
        "example": {
          "usage": {
            "summary": "Usage Example",
            "value": "To collect data from multiple pages linked from `https://wsa-test.vercel.app`, make a GET request to `http://localhost:5000/getAll?url=https://wsa-test.vercel.app`."
          }
        }
      }
    },
    "/get-all-from-script": {
      "get": {
        "summary": "Scrape multiple pages using a script.",
        "description": "Use this endpoint to collect data from multiple web pages linked from the specified URL",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "schema": {
              "type": "string"
            },
            "description": "The URL of the web page to start scraping from."
          }
        ],
        "responses": {
          "200": {
            "description": "Data retrieved successfully."
          }
        }
      },
      "example": {
        "usage": {
          "summary": "Usage Example",
          "value": "To collect data from multiple pages linked from `https://wsa-test.vercel.app`, make a GET request to `http://localhost:5000/get-all-from-script?url=https://wsa-test.vercel.app`."
        }
      }
    }
  },
  "components": {},
  "tags": []
}